pyproject.toml
[project]
name = "intel2sigma"
version = "0.0.1"
description = "Turn observed malware behaviors into shareable Sigma rules and SIEM queries."
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "davidsharp.io" }]
keywords = ["sigma", "detection", "siem", "malware", "cti"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security",
]

dependencies = [
    # Web + CLI
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "jinja2>=3.1,<4.0",
    "typer>=0.12,<1.0",
    # Core
    "pydantic>=2.7,<3.0",
    "ruamel.yaml>=0.18,<1.0",
    "pygments>=2.18,<3.0",
    # Sigma engine
    "pysigma>=0.11,<1.0",
    # Conversion backends
    "pysigma-backend-kusto>=0.4,<2.0",
    "pysigma-backend-splunk>=1.0,<3.0",
    "pysigma-backend-elasticsearch>=1.1,<3.0",
    "pysigma-backend-crowdstrike>=1.0,<4.0",
    # Pipelines (microsoft-xdr pipelines ship inside pysigma-backend-kusto;
    # splunk/windows/crowdstrike pipelines are separate packages)
    "pysigma-pipeline-sysmon>=1.0,<3.0",
    "pysigma-pipeline-windows>=1.0,<3.0",
    "pysigma-pipeline-crowdstrike>=1.0,<3.0",
]

[project.scripts]
intel2sigma = "intel2sigma.cli.main:app"

[dependency-groups]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-cov>=5.0,<7.0",
    "mypy>=1.11,<2.0",
    "ruff>=0.6,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["intel2sigma"]

# --- Ruff -----------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py314"
extend-exclude = [".venv", "dist", "build"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "UP",   # pyupgrade
    "SIM",  # simplify
    "RUF",  # ruff-specific
    "C4",   # comprehensions
    "N",    # pep8-naming
    "PL",   # pylint
    "TRY",  # tryceratops
    "PTH",  # flake8-use-pathlib
    "A",    # builtins shadowing
]
ignore = [
    "PLR2004",  # magic-value-comparison — too noisy in tests
    "PLR0913",  # too-many-arguments — pydantic models legitimately have many fields
    "TRY003",   # long exception messages — we prefer explicit user-facing messages
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLR", "N802", "N803"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# --- Mypy -----------------------------------------------------------------

[tool.mypy]
python_version = "3.14"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_any_generics = true
disallow_untyped_decorators = true
no_implicit_reexport = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
# pySigma ships without a py.typed marker for some backends.
module = [
    "sigma.*",
    "ruamel.*",
]
ignore_missing_imports = true

# --- Pytest ---------------------------------------------------------------

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (e.g. SigmaHQ corpus parse); deselect with -m 'not slow'",
]

# --- Coverage -------------------------------------------------------------

[tool.coverage.run]
branch = true
source = ["intel2sigma"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
